В какой-то из инструкций-видеороликов говорилось, что пока мы создаём "тесты" для проверки в мейне. Создала для задачи тестовый мейн. Убрала всё из основного мейна во избежание конфликтов
